requirements.txt
anthropic
edgartools
hyperbrowser
python-dotenv
